First real commit
First real commit of working dual screens and updated install files.

+++ userscripts/game-end/marquee-end.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
+echo "Script called at: $(date)" >> $LOGFILE
 
 # DEBUG: Log all arguments to a file
+#Set to true to persist log file 
+DEBUG=false
 LOGFILE="/rcade/share/userscripts/game-end/marquee-end-debug.log"
-rm -f "$LOGFILE"
 
+
+if [ "$DEBUG" != "true" ]; then
+  rm -f "$LOGFILE"
+fi
 touch $LOGFILE
 
 echo "Sending command: CLOSE" >> $LOGFILE
 
-python3 /rcade/share/userscripts/simpleClient.py "CLOSE"  >> $LOGFILE 2>&1
-#echo "Sending command: OPEN all emulation_station" >> $LOGFILE
+echo "CLOSE" > /tmp/marquee-daemon.pipe
 
-#python3 /rcade/share/userscripts/simpleClient.py "OPEN all emulation_station"  >> $LOGFILE 2>&1
 
 echo "Script completed" >> $LOGFILE
