+++ userscripts/reboot/marquee-reboot.sh
@@ -2,8 +2,14 @@
 #!/bin/bash
 
 # DEBUG: Log all arguments to a file
+#Set to true to persist log file 
+DEBUG=false
+
 LOGFILE="/rcade/share/userscripts/reboot/marquee-reboot-debug.log"
-rm -f "$LOGFILE"
+
+if [ "$DEBUG" != "true" ]; then
+  rm -f "$LOGFILE"
+fi
 touch &LOGFILE
 echo "Initiating arcade reboot sequence..." >> $LOGFILE
 
@@ -13,6 +19,7 @@ echo "Initiating arcade reboot sequence..." >> $LOGFILE
 
 # Send shutdown command to marquee display
 echo "Sending shutdown to marquee display..." >> $LOGFILE
-python3 /rcade/share/userscripts/simpleClient.py "REBOOT"
+
+echo "REBOOT" > /tmp/marquee-daemon.pipe
 
 echo "Rebooting sequence complete">> $LOGFILE
