+++ userscripts/screensaver-stop/marquee-screensaver-stop.sh
@@ -2,17 +2,22 @@
 #!/bin/bash
 
 # DEBUG: Log all arguments to a file
+
+#Set to true to persist log file 
+DEBUG=false
+
 LOGFILE="/rcade/share/userscripts/screensaver-stop/marquee-screensaver-stop-debug.log"
-rm -f "$LOGFILE"
+if [ "$DEBUG" != "true" ]; then
+  rm -f "$LOGFILE"
+fi
 touch &LOGFILE
+
+
 echo "Stopping marquee screensaver..." >> $LOGFILE
 
-# Notify Home Assistant
-#echo "Notifying Home Assistant..." >> $LOGFILE
-#python3 /rcade/share/userscripts/ha-shutdown.py shutdown
-
-# Send shutdown command to marquee display
+# Send screensaver stop command to marquee display
 echo "Sending screensaver stop command to marquee display..." >> $LOGFILE
-python3 /rcade/share/userscripts/simpleClient.py "SCREENSAVER-STOP"
+
+echo "SCREENSAVER-STOP" > /tmp/marquee-daemon.pipe
 
 echo "Screensaver stop complete">> $LOGFILE
